# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MemBrain API
  version: 1.0.0
  description: Strategic Collective Forgetting Agent — REST API specification
servers:
  - url: http://localhost:3001
    description: Development server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: User registered successfully
  /api/auth/login:
    post:
      summary: Login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
...